converting to stack view
